local-draw-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
Pillow==10.4.0
diffusers==0.30.3
transformers==4.44.2
accelerate==0.34.2
safetensors==0.4.5
huggingface-hub==0.24.6
